[#66550] inference_server: Fix error messages in 'reset_runtime_alloc'

+++ lib/kenning_inference_lib/core/callbacks.c
@@ -266,7 +266,7 @@ status_t runtime_callback(message_hdr_t *hdr, resp_message_t *resp)
         struct llext *p_llext = llext_by_name("runtime");
         if (IS_VALID_POINTER(p_llext))
         {
-            LOG_ERR("LLEXT already exists (%x). This should never happen", p_llext);
+            LOG_ERR("LLEXT already exists (%x). This should never happen", (unsigned int)p_llext);
             return CALLBACKS_STATUS_ERROR;
         }
